gita-agent/requirements.txt
# Core Framework
langchain>=0.2.0
langchain-community>=0.2.0
langchain-anthropic>=0.1.0

# Vector Database
chromadb>=0.4.0

# Embeddings
sentence-transformers>=2.2.0

# LLM Providers
anthropic>=0.25.0

# Frontend
streamlit>=1.30.0

# Environment
python-dotenv>=1.0.0

# Utilities
tiktoken>=0.5.0